% Generated by roxygen2 (4.0.2): do not edit by hand
\name{branches}
\alias{branches}
\title{Color branches of a polar dendrogram}
\usage{
branches(x, fill = FALSE, lwd, sector = c(0, 2 * pi), col, alpha = 1,
  lightness, ...)
}
\arguments{
\item{x}{\code{\link{polar.clust}} object.}

\item{fill}{Annotation style, see the details of this function.}

\item{lwd}{Line width, should be pretty large to be clearly visible.}

\item{sector}{See \code{\link{plot.polar.clust}}.}

\item{col}{Branch colors. If omitted it will be taken from \code{x} (if
it has first been specified with \code{\link{colors<-}}).}

\item{alpha}{Opacity values (0 is transparent, 1 is opaque).}

\item{lightness}{Color lightness (0 is the original color, 1 is white, and
values in between are interpolated).}

\item{...}{Sent to \code{\link{polygon}} or \code{\link{lines}}.}
}
\description{
This function can either draw a line in the margin outside each branch
(\code{fill=FALSE}) or color the area underneath a branch (\code{fill=TRUE}).
}
\examples{
my.data <- iris[-5] + .5*matrix(rnorm(nrow(iris)*(ncol(iris)-1)), nrow(iris))
cl <- hclust(dist(my.data))
x <- polar.clust(cl)

plot(x)

# Color according to class
labels(x) <- sprintf("\%s \%i", iris$Species, 1:150)
colors(x) <- iris$Species
plot(x, expand=1.2, label.size=.3, las=2)

# Color according to branch
colors(x) <- c(brewer.pal(6, "Set1"))
labels(x) <- NULL # Clear the current labels
labels(x) <- paste("Branch", 1:6)
plot(x, las=1)

# Annotate manually
labels(x) <- NULL
x <- label.branches(x)
plot(x, las=1)

# Manual compilation of the plot
s <- c(.4,2.2)*pi
colors(x) <- NULL
plot(x, type="n", labels=FALSE, axes=FALSE, expand=1.1)
branches(x, k=6, sector=s, fill=TRUE, lwd=10, lightness=.3)
lines(x, sector=s)
points(x, pch=19, sector=s)
legend("topright", paste("Branch", 1:6), fill=brewer.pal(6, "Set1"), bg="white")

}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

