% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polar.clust.R
\name{colors<-}
\alias{colors<-}
\alias{colors<-.polar.clust}
\alias{labels<-}
\alias{labels<-.polar.clust}
\title{colors-set or labels to a polar dendrogram}
\usage{
colors(object) <- value

\method{colors}{polar.clust}(object) <- value

labels(object) <- value

\method{labels}{polar.clust}(object) <- value
}
\arguments{
\item{object}{The \code{\link{polar.clust}} object.}

\item{value}{The colors or labels of the cluster leaves. If \code{value}
has the same length as the number of leaves it corresponds to the leaves
(ordered as in the original dataset). Otherwise \code{value} corresponds
to branches (indexed by \code{\link{cutree.polar.clust}}).}
}
\value{
The modified cluster object
}
\description{
colors-set or labels to a polar dendrogram
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

