% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cutree.polar.clust}
\alias{cutree.polar.clust}
\title{Cut a tree inte groups}
\usage{
cutree.polar.clust(tree, k, h)
}
\arguments{
\item{tree}{\code{\link{polar.clust}} object.}

\item{k}{Number of groups.}

\item{h}{Height at which to cut the tree.}
}
\description{
A wrapper that allows cutting polar dendrograms with \code{\link{cutree}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
cutree
}

