% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.branch.index}
\alias{get.branch.index}
\title{Extracting line numbers of desired branches}
\usage{
get.branch.index(x, k, h)
}
\arguments{
\item{x}{\code{\link{polar.clust}} object.}

\item{k}{Number of branches to split \code{x} into.}

\item{h}{Height at which to split \code{x}.}
}
\description{
Complement to \code{\link{cutree.polar.clust}}.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

