% Generated by roxygen2 (4.0.2): do not edit by hand
\name{label.branches}
\alias{label.branches}
\title{Annotate a polar dendrogram}
\usage{
label.branches(x)
}
\arguments{
\item{x}{\code{\link{polar.clust}} object.}
}
\value{
A modified \code{\link{polar.clust}} object.
}
\description{
Annotate a polar dendrogram
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

