% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/branches.R
\name{locate.branches}
\alias{locate.branches}
\title{Locate branches graphically}
\usage{
locate.branches(x)
}
\arguments{
\item{x}{\code{\link{polar.clust}} object.}
}
\value{
A data.table with relevant attributes.
}
\description{
Locate branches graphically
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

