% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polar.clust.R
\name{order.polar.clust}
\alias{order.polar.clust}
\title{Sample order in clustering}
\usage{
order.polar.clust(x)
}
\arguments{
\item{x}{\code{\link{polar.clust}} object.}
}
\value{
The clockwise order in which the samples appear in a polar clustering.
}
\description{
Sample order in clustering
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

