% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot.polar.clust}
\alias{lines.polar.clust}
\alias{plot.polar.clust}
\alias{points.polar.clust}
\alias{text.polar.clust}
\title{Plot functions for polar dendrograms}
\usage{
\method{plot}{polar.clust}(x, type = c("b", "l", "p", "n"), sector,
  fill = TRUE, labels, label.col, label.size = 1, las = 2, expand,
  axes = TRUE, ..., bg.col, grid.col = "gray90")

\method{lines}{polar.clust}(x, lwd = c(1, 10), col, sector = c(0, 2 * pi),
  ...)

\method{text}{polar.clust}(x, r, sector = c(0, 2 * pi), cex = par("cex"),
  col, las = par("las"), ...)

\method{points}{polar.clust}(x, sector = c(0, 2 * pi), col, ...)
}
\arguments{
\item{x}{\code{\link{polar.clust}} object.}

\item{type}{Plot type, see \code{\link{plot}}.
you add whatever layers you want manually.}

\item{sector}{The sector of the plot area to use, e.g. set
\code{sector=c(1/2, 2)*pi} to leave one quadrant empty, giving room to put
a \code{\link{legend}}.}

\item{fill}{Whether to fill the area under labelled branches, see \code{\link{branches}}.}

\item{labels}{Whether to plot the leaves' labels.}

\item{label.col}{Whether to colour labels by leaf colour.}

\item{label.size}{Relative text size of labels.}

\item{las}{Label orientation, see \code{\link{par}}.}

\item{expand}{Controls how much white space should surround the plot.
Use \code{expand=1} for no white space and larger numbers for more.}

\item{axes}{Whether to plot axes or not.}

\item{...}{Sent to \code{\link{plot}}.}

\item{bg.col}{Background colour.}

\item{grid.col}{Radial grid color.}

\item{lwd}{Minimum and maximum branch widths.}

\item{col}{Colour.}

\item{r}{Radius at which to plot the labels.}

\item{cex}{Relative size, see \code{\link{par}} for details.}

\item{lines}{Whether to plot the trees branches.}

\item{points}{Whether to plot the trees leaves.}
}
\value{
Nothing, produces a plot.
}
\description{
Class specific methods for class \code{\link{polar.clust}}. If you are not
happy with the looks of the plot you can build it up manually by calling it
with \code{type="n"} and add the annotation functions.
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

