% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polar.clust.R
\name{polar.clust}
\alias{polar.clust}
\title{Convert a hclust object to a polar dendrogram}
\usage{
polar.clust(x, labels = sprintf("#\%i", 1:length(x$order)), col,
  padding = 0.01, hang = 0.1 * max(x$height))
}
\arguments{
\item{x}{\code{\link{hclust}} object.}

\item{labels}{Chraracter vector of leaf labels.}

\item{col}{Vector of leaf colours.}

\item{padding}{Spacing between branches.}

\item{hang}{Branch lengths of the leaves.}
}
\value{
An object of class \code{polar.clust}
}
\description{
Any \code{\link{hclust}} object can be converted to a polar dendrogram
using this function and then plotted with the standard graphics functions.
}
\examples{
my.data <- iris[-5] + .5*matrix(rnorm(nrow(iris)*(ncol(iris)-1)), nrow(iris))
cl <- hclust(dist(my.data))
x <- polar.clust(cl)

plot(x)

# Color according to class
labels(x) <- sprintf("\%s \%i", iris$Species, 1:150)
colors(x) <- iris$Species
plot(x, expand=1.2, label.size=.3, las=2)

# Color according to branch
colors(x) <- c(brewer.pal(6, "Set1"))
labels(x) <- NULL # Clear the current labels
labels(x) <- paste("Branch", 1:6)
plot(x, las=1)

# Manual compilation of the plot
s <- c(.4,2.2)*pi
colors(x) <- NULL
plot(x, type="n", labels=FALSE, axes=FALSE, expand=1.1)
branches(x, k=6, sector=s, fill=TRUE, lwd=10, lightness=.3)
lines(x, sector=s)
points(x, pch=19, sector=s)
legend("topright", paste("Branch", 1:6), fill=brewer.pal(6, "Set1"), bg="white")

\dontrun{# Annotate manually
labels(x) <- NULL
x <- label.branches(x)
plot(x, las=1)
}
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}
\seealso{
plot.polar.clust
}

