% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/branches.R
\name{label.branches}
\alias{label.branches}
\title{Annotate a polar dendrogram}
\usage{
label.branches(x, labels, col, redraw = TRUE)
}
\arguments{
\item{x}{\code{\link{polar.clust}} object.}
}
\value{
A modified \code{\link{polar.clust}} object.
}
\description{
Annotate a polar dendrogram
}
\author{
Christofer \enc{Bäcklin}{Backlin}
}

